fix type
Former-commit-id: 2736fb8335f459608cd9bbbcbfc668f8965701a7

+++ oneflow/python/model/resnet50.py
@@ -309,7 +309,7 @@ def mock_dataset():
             "images": np.random.randn(*images_shape).astype(np.float32),
             "labels": np.random.randint(
                 low=0, high=2, size=labels_shape
-            ).astype(np.float32),
+            ).astype(np.int32),
         },
     }
